+++ src/Gui/ToolBarManager.cpp
@@ -21,18 +21,14 @@
  ***************************************************************************/
 
 #include "PreCompiled.h"
-#include "Selection.h"
-#include <boost/signals2/connection.hpp>
 #ifndef _PreComp_
-# include <QAction>
-# include <QApplication>
-# include <QHBoxLayout>
-# include <QMenuBar>
-# include <QMouseEvent>
-# include <QPointer>
-# include <QStatusBar>
-# include <QToolBar>
-# include <QToolButton>
+#include <QAction>
+#include <QApplication>
+#include <QHBoxLayout>
+#include <QMenuBar>
+#include <QMouseEvent>
+#include <QStatusBar>
+#include <QToolButton>
 #endif
 
 #include <boost/algorithm/string/predicate.hpp>
@@ -40,6 +36,7 @@
 #include <Base/Tools.h>
 
 #include "ToolBarManager.h"
+#include "ToolBarAreaWidget.h"
 #include "Application.h"
 #include "Command.h"
 #include "MainWindow.h"
@@ -162,113 +159,6 @@ QList<ToolBarItem*> ToolBarItem::getItems() const
 {
     return _items;
 }
-
-// -----------------------------------------------------------
-ToolBarAreaWidget::ToolBarAreaWidget(QWidget* parent,
-                                          ToolBarArea area,
-                                          const ParameterGrp::handle& hParam,
-                                          boost::signals2::scoped_connection& conn,
-                                          QTimer* timer)
-    : QWidget(parent)
-    , _sizingTimer(timer)
-    , _hParam(hParam)
-    , _conn(conn)
-    , _area(area)
-{
-    _layout = new QHBoxLayout(this);
-    _layout->setContentsMargins(QMargins());
-}
-
-void ToolBarAreaWidget::addWidget(QWidget* widget)
-{
-    // if widget already exist don't do anything
-    if (_layout->indexOf(widget) >= 0) {
-        return;
-    }
-
-    _layout->addWidget(widget);
-    adjustParent();
-
-    QString name = widget->objectName();
-
-    if (!name.isEmpty()) {
-        Base::ConnectionBlocker block(_conn);
-        _hParam->SetInt(widget->objectName().toUtf8().constData(), _layout->count() - 1);
-    }
-}
-
-void ToolBarAreaWidget::insertWidget(int index, QWidget* widget)
-{
-    int currentIndex = _layout->indexOf(widget);
-
-    // we are inserting widget at the same place, this is no-op
-    if (currentIndex == index) {
-        return;
-    }
-
-    // widget already exists in the area, we need to first remove it and then recreate
-    if (currentIndex > 0) {
-        _layout->removeWidget(widget);
-    }
-
-    _layout->insertWidget(index, widget);
-
-    adjustParent();
-    saveState();
-}
-
-void ToolBarAreaWidget::removeWidget(QWidget* widget)
-{
-    _layout->removeWidget(widget);
-
-    QString name = widget->objectName();
-    if (!name.isEmpty()) {
-        Base::ConnectionBlocker block(_conn);
-        _hParam->RemoveInt(name.toUtf8().constData());
-    }
-
-    adjustParent();
-}
-
-void ToolBarAreaWidget::adjustParent()
-{
-    if (_sizingTimer) {
-        _sizingTimer->start(10);
-    }
-}
-
-void ToolBarAreaWidget::saveState()
-{
-    Base::ConnectionBlocker block(_conn);
-
-    for (auto &v : _hParam->GetIntMap()) {
-        _hParam->RemoveInt(v.first.c_str());
-    }
-
-    foreachToolBar([this](QToolBar *toolbar, int idx, ToolBarAreaWidget*) {
-        _hParam->SetInt(toolbar->objectName().toUtf8().constData(), idx);
-    });
-}
-
-void ToolBarAreaWidget::restoreState(const std::map<int, QToolBar*>& toolbars)
-{
-    for (const auto &[index, toolbar] : toolbars) {
-        bool visible = toolbar->isVisible();
-        getMainWindow()->removeToolBar(toolbar);
-        toolbar->setOrientation(Qt::Horizontal);
-        insertWidget(index, toolbar);
-        toolbar->setVisible(visible);
-    }
-
-    for (const auto &[name, visible] : _hParam->GetBoolMap()) {
-        auto widget = findChild<QWidget*>(QString::fromUtf8(name.c_str()));
-
-        if (widget) {
-            widget->setVisible(visible);
-        }
-    }
-}
-
 
 // -----------------------------------------------------------
