+++ src/Gui/ToolBarManager.h
@@ -29,7 +29,6 @@
 
 #include <QStringList>
 #include <QTimer>
-#include <QLayout>
 #include <QToolBar>
 #include <QPointer>
 
@@ -43,84 +42,8 @@ class QMouseEvent;
 
 namespace Gui {
 
-// Qt treats area as Flag so in theory toolbar could be in multiple areas at once.
-// We don't do that here so simple enum should suffice.
-enum class ToolBarArea {
-    NoToolBarArea,
-    LeftToolBarArea,
-    RightToolBarArea,
-    TopToolBarArea,
-    BottomToolBarArea,
-    LeftMenuToolBarArea,
-    RightMenuToolBarArea,
-    StatusBarToolBarArea,
-};
-
-class ToolBarAreaWidget : public QWidget
-{
-    Q_OBJECT
-    using inherited = QWidget;
-
-public:
-    ToolBarAreaWidget(QWidget *parent,
-                ToolBarArea area,
-                const ParameterGrp::handle& hParam,
-                boost::signals2::scoped_connection &conn,
-                QTimer *timer = nullptr);
-
-    void addWidget(QWidget *widget);
-    void insertWidget(int index, QWidget *widget);
-    void removeWidget(QWidget *widget);
-
-    void adjustParent();
-
-    QWidget *widgetAt(int index) const
-    {
-        auto item = _layout->itemAt(index);
-
-        return item ? item->widget() : nullptr;
-    }
-
-    int count() const
-    {
-        return _layout->count();
-    }
-
-    int indexOf(QWidget *widget) const
-    {
-        return _layout->indexOf(widget);
-    }
-
-    ToolBarArea area() const
-    {
-        return _area;
-    }
-
-    template<class FuncT>
-    void foreachToolBar(FuncT &&func)
-    {
-        for (int i = 0, count = _layout->count(); i < count; ++i) {
-            auto toolbar = qobject_cast<QToolBar*>(widgetAt(i));
-
-            if (!toolbar || toolbar->objectName().isEmpty()
-                         || toolbar->objectName().startsWith(QStringLiteral("*"))) {
-                continue;
-            }
-
-            func(toolbar, i, this);
-        }
-    }
-
-    void saveState();
-    void restoreState(const std::map<int, QToolBar*> &toolbars);
-
-private:
-    QHBoxLayout *_layout;
-    QPointer<QTimer> _sizingTimer;
-    ParameterGrp::handle _hParam;
-    boost::signals2::scoped_connection &_conn;
-    ToolBarArea _area;
-};
+class ToolBarAreaWidget;
+enum class ToolBarArea;
 
 class GuiExport ToolBarItem
 {
